refactor: rename infrastructure/ → integrations/dynatrace/ + relocate core modules
Restructure container/infrastructure/ which conflated DT-specific
integration code with generic K8s operations:

Moved to container/integrations/dynatrace/:
  dt_data_models.py    → data_models.py
  dt_enrichment.py     → enrichment.py
  dt_mcp_client.py     → mcp_client.py
  dt_query_strategies.py → query_strategies.py

Moved to container/ (K8s-native operations, not integrations):
  resource_applier.py
  enforcement.py

Adding a new observability integration (Datadog, Falco, etc.) is now:
  1. Create container/integrations/<provider>/
  2. Implement EnrichmentProvider protocol
  3. Add to _create_enrichment_provider() factory in cli/generate.py

387 tests, zero lint errors.
