+++ tests/test_dt_query_strategies.py
@@ -17,12 +17,12 @@ from unittest.mock import AsyncMock, MagicMock, patch
 
 import pytest
 
-from kimera.container.infrastructure.dt_data_models import (
+from kimera.container.integrations.dynatrace.data_models import (
     DtContext,
     KspmFinding,
     SmartscapeEdge,
 )
-from kimera.container.infrastructure.dt_query_strategies import (
+from kimera.container.integrations.dynatrace.query_strategies import (
     EXPLOIT_KSPM_FILTERS,
     EXPLOIT_SMARTSCAPE_SCOPE,
     DavisStrategy,
@@ -269,7 +269,7 @@ class TestLlmQueryStrategy:
         result = LlmQueryStrategy._parse_query_response(raw)
         assert result == []
 
-    @patch("kimera.container.infrastructure.dt_query_strategies.LlmQueryStrategy._generate_queries")
+    @patch("kimera.container.integrations.dynatrace.query_strategies.LlmQueryStrategy._generate_queries")
     def test_fetch_pipeline(self, mock_generate: MagicMock) -> None:
         mock_generate.return_value = [
             {"purpose": "kspm", "query": "fetch security.events, from:-24h"},
@@ -282,7 +282,7 @@ class TestLlmQueryStrategy:
         assert ctx.strategy_name == "llm-query"
         assert len(ctx.queries_executed) == 1
 
-    @patch("kimera.container.infrastructure.dt_query_strategies.LlmQueryStrategy._generate_queries")
+    @patch("kimera.container.integrations.dynatrace.query_strategies.LlmQueryStrategy._generate_queries")
     def test_fetch_skips_invalid_queries(self, mock_generate: MagicMock) -> None:
         mock_generate.return_value = [
             {"purpose": "bad", "query": "SELECT * FROM foo"},
@@ -293,7 +293,7 @@ class TestLlmQueryStrategy:
         ctx = _run(strategy.fetch(client, "all", "unguard", ""))
         assert len(ctx.queries_executed) == 1  # Only valid query executed
 
-    @patch("kimera.container.infrastructure.dt_query_strategies.LlmQueryStrategy._generate_queries")
+    @patch("kimera.container.integrations.dynatrace.query_strategies.LlmQueryStrategy._generate_queries")
     def test_fetch_handles_no_queries(self, mock_generate: MagicMock) -> None:
         mock_generate.return_value = []
         client = _mock_mcp_client()
