+++ tests/test_exploit_generation.py
@@ -19,8 +19,8 @@ import pytest
 
 from kimera.container.core.k8s_client import K8sClient
 from kimera.container.core.logger import SecurityLogger
-from kimera.container.infrastructure.resource_applier import ResourceApplier
 from kimera.container.remediations.generator import LLMRemediationGenerator
+from kimera.container.resource_applier import ResourceApplier
 
 
 def _create_mock_k8s_client() -> tuple[K8sClient, MagicMock]:
